requirements.txt
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.12.2
django-nose==1.4.7
gunicorn==20.0.4
pandas==1.2.3
phillydb==0.0.7
psycopg2==2.8.6
requests==2.25.1
urllib3==1.26.4
whitenoise==5.2.0

# testing
pytest==6.2.2
pytest-cov==2.11.1
pytest-xdist==2.2.1
pytest-benchmark==3.2.3
pytest-profiling==1.7.0
pytest-django==4.1.0
coverage[toml]==5.5
